Refactor dependencies for API Services.
