+++ src/AppBundle/Service/LinodeHostService.php
@@ -8,6 +8,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use AppBundle\Model\DataCenter;
 use AppBundle\Enum\PaymentTermEnum;
 use AppBundle\Model\LinodePlan;
+use Linode\LinodeApi;
 
 /**
  * Class LinodeHostService
@@ -23,9 +24,9 @@ class LinodeHostService
      *
      * @param LinodeApiService
      */
-    public function __construct(LinodeApiService $api)
+    public function __construct(LinodeApi $api)
     {
-        $this->api = $api->get('LinodeApi');
+        $this->api = $api;
     }
 
     public function getLinodes($linodeId = null)
